+++ src/lib.rs
@@ -487,6 +487,7 @@ pub mod afv;
 pub mod ans;
 pub mod begabrac;
 pub mod bitreader;
+pub mod block_context_resolver;
 pub mod chroma_from_luma;
 pub mod coeff_order;
 pub mod container;